+++ lib/bloc_after_effect.dart
@@ -1,2 +1,3 @@
+export 'src/bloc_effect_consumer.dart';
 export 'src/bloc_effect_listener.dart';
 export 'src/effect_bloc.dart';